test: rename concurrency specs to race specs and assert invariants
